+++ astropy/modeling/models.py
@@ -14,6 +14,7 @@ from .functional_models import *
 from .powerlaws import *
 from .tabular import *
 from .blackbody import BlackBody1D
+from . import math_functions as math
 
 
 """